Refactor: move torch helpers to simpler_setup.torch_interop
Move ``make_tensor_arg`` and ``torch_dtype_to_datatype`` (plus the
private ``_TORCH_DTYPE_MAP`` / ``_ensure_torch_map``) out of
``simpler.task_interface`` and into a new ``simpler_setup/torch_interop.py``.
The ``simpler`` package is the stable runtime user API and nothing in
``python/simpler/`` actually uses these helpers — all callers are in
``simpler_setup``, examples, or tests. ``simpler_setup`` already hosts
torch-using modules (goldens, scene_test). Consolidating keeps the
runtime API torch-free end to end.

- Add ``simpler_setup/torch_interop.py`` as the canonical home. The
  ``hasattr(torch, "uint16/uint32")`` guard added in #615 is dropped:
  this module is only imported by callers who opt into torch, and
  torch >= 2.3 (which introduced those dtypes) is now the supported
  minimum for that opt-in path.
- Re-export ``make_tensor_arg`` and ``torch_dtype_to_datatype`` from
  ``simpler_setup/__init__.py`` so callers can ``from simpler_setup
  import make_tensor_arg``.
- Remove the old definitions from ``simpler.task_interface`` outright.
  All in-repo callers are migrated in the same commit, so a deprecation
  shim would protect no one — prefer an honest ImportError to a silent
  forward. ``scalar_to_uint64`` (not torch-specific) stays where it is.
- Migrate in-repo callers to the new location: ``simpler_setup/
  scene_test.py``, ``examples/workers/l3/multi_chip_dispatch/main.py``,
  and the three ``tests/st/a2a3/.../test_l3_*.py`` scene tests.
- Update unit tests: move torch-aware tests onto the new import and
  add a ``uint16``/``uint32`` torch-dtype case now that the guard is
  gone.

Closes #617.
