+++ tests/st/a2a3/tensormap_and_ringbuffer/test_l3_dependency.py
@@ -16,9 +16,9 @@ SubWorker reads result produced by ChipWorker.
 
 import torch
 from simpler.task_interface import ArgDirection as D
-from simpler.task_interface import TaskArgs, TensorArgType, make_tensor_arg
+from simpler.task_interface import TaskArgs, TensorArgType
 
-from simpler_setup import SceneTestCase, TaskArgsBuilder, Tensor, scene_test
+from simpler_setup import SceneTestCase, TaskArgsBuilder, Tensor, make_tensor_arg, scene_test
 from simpler_setup.scene_test import _build_l3_task_args
 
 KERNELS_BASE = "../../../../examples/a2a3/tensormap_and_ringbuffer/vector_example/kernels"
