+++ tests/st/a2a3/tensormap_and_ringbuffer/test_l3_group.py
@@ -17,9 +17,9 @@ group completion aggregation, downstream dependency waits for group.
 
 import torch
 from simpler.task_interface import ArgDirection as D
-from simpler.task_interface import TaskArgs, TensorArgType, make_tensor_arg
+from simpler.task_interface import TaskArgs, TensorArgType
 
-from simpler_setup import SceneTestCase, TaskArgsBuilder, Tensor, scene_test
+from simpler_setup import SceneTestCase, TaskArgsBuilder, Tensor, make_tensor_arg, scene_test
 
 KERNELS_BASE = "../../../../examples/a2a3/tensormap_and_ringbuffer/vector_example/kernels"
